edgefirst_lidarpub::lidar

Trait LidarFrameWriter

pub trait LidarFrameWriter: LidarFrame {
    // Required methods
    fn set_timestamp(&mut self, ts: u64);
    fn set_frame_id(&mut self, id: u32);
    fn push(
        &mut self,
        x: f32,
        y: f32,
        z: f32,
        intensity: u8,
        range: f32,
    ) -> bool;
    fn set_len(&mut self, len: usize);
    fn buffers_mut(
        &mut self,
    ) -> (&mut [f32], &mut [f32], &mut [f32], &mut [u8], &mut [f32]);
}

Internal trait for drivers to write into frames.

This trait is not intended for public use - clients should only use the read-only LidarFrame trait methods.

Required Methods

fn set_timestamp(&mut self, ts: u64)

Set the frame timestamp in nanoseconds.

fn set_frame_id(&mut self, id: u32)

Set the frame sequence ID.

fn push(&mut self, x: f32, y: f32, z: f32, intensity: u8, range: f32) -> bool

Push a point to the frame.

Returns true if the point was added, false if at capacity.

fn set_len(&mut self, len: usize)

Set the valid length of the frame.

fn buffers_mut( &mut self, ) -> (&mut [f32], &mut [f32], &mut [f32], &mut [u8], &mut [f32])

Get mutable access to all buffers at once for SIMD writes.

Returns (x, y, z, intensity, range) slices. This method allows borrowing all buffers simultaneously, which is necessary for efficient SIMD processing where we need to write to multiple arrays in a single pass.